# server/src/docs/swagger.yaml
openapi: 3.1.0
info:
  title: Hypertube API
  version: 0.1.0
  description: API for Hypertube streaming platform

# Movie Interaction endpoints are defined in swagger.interactions.yaml
# To view them, check the swagger.interactions.yaml file in the same directory
paths:
...
  /api/v1/auth/refresh-token:
    post:
      summary: Refresh access token using refresh token
      tags:
        - Authentication
      parameters:
        - name: refresh_token
          in: cookie
          description: Refresh token issued during login
          required: true
          schema:
            type: string
      responses:
        '200':
          description: New access token issued - sets httpOnly cookie
          headers:
            Set-Cookie:
              schema:
                type: string
                example: accessToken=eyJhbGc...; HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=3600
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Token refreshed successfully
        '401':
          description: Invalid, expired, or missing refresh token
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: fail
                  message:
                    type: string
                    example: No refresh token provided
                  path:
                    type: string
                    example: /api/v1/auth/refresh-token
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: JWT access token stored in httpOnly cookie (set after login or OAuth callback)
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from login or OAuth callback (alternative to cookie auth)